+++ src/components/ZRange/components/ZRangeLabel/index.tsx
@@ -5,7 +5,7 @@ import styles from "../../styles/ZRange.module.scss";
 const ZRangeLabel = (props: ZRangeLabelProps) => {
   const { label, labelClassName } = props;
 
-  const labelClasses = classNames(styles["z-range__label"], labelClassName);
+  const labelClasses = classNames(styles["zr__label"], labelClassName);
 
   return <div className={labelClasses}>{label}</div>;
 };