+++ src/index.ts
@@ -1,7 +1,14 @@
 export { default as ZDrop } from "./components/ZDrop";
-export { ZDropField } from "./components/ZDrop/integrations/react-hook-form/ZDropField";
-export { default as ZDropButton } from "./components/ZDropButton";
-export { default as ZRange } from "./components/ZRange";
 export type * from "./components/ZDrop/types/zDropTypes";
+
+export { default as ZDropField } from "./components/ZDrop/integrations/react-hook-form/ZDropField";
+export type { ZDropFieldProps } from "./components/ZDrop/integrations/react-hook-form/ZDropField/types/zDropFieldTypes";
+
+export { default as ZDropButton } from "./components/ZDropButton";
 export type * from "./components/ZDropButton/types/zDropButtonTypes";
+
+export { default as ZRange } from "./components/ZRange";
 export type * from "./components/ZRange/types/zRangeTypes";
+
+export { default as ZRangeField } from "./components/ZRange/integrations/react-hook-form/ZRangeField";
+export type { ZRangeFieldProps } from "./components/ZRange/integrations/react-hook-form/ZRangeField/types/zRangeFieldTypes";