+++ src/stories/ZComponents/Integrations/ReactHookForm/SubmittedMessage/index.tsx
@@ -1,0 +1,11 @@
+interface Props {
+  className?: string;
+}
+
+const SubmittedMessage = (props: Props) => {
+  const { className } = props;
+
+  return <p className={className}>Form submitted!</p>;
+};
+
+export default SubmittedMessage;