+++ src/stories/ZComponents/Integrations/ReactHookForm/ValuePreview/index.tsx
@@ -1,0 +1,17 @@
+interface Props {
+  className?: string;
+  values?: any;
+  preparePreviewValue: (value: any) => any;
+}
+
+const ValuePreview = (props: Props) => {
+  const { className, values, preparePreviewValue } = props;
+
+  return (
+    <div className={className}>
+      <pre>{JSON.stringify(preparePreviewValue(values), null, 2)}</pre>
+    </div>
+  );
+};
+
+export default ValuePreview;